Download update according to app type (portable/perUser)

+++ src/Infrastructure/AppWindow.cs
@@ -145,19 +145,23 @@
             // This behavior is enabled by default to allow toast notifications because, without a valid shortcut installed, Photino cannot raise a toast notification from a desktop app.
             // If the user has chosen to activate the application shortcut during app installation, this results in a duplicate of the application shortcut in the Windows start menu.
             // The issue has been reported on GitHub, meanwhile let's get rid of the shortcut created by Photino https://github.com/tryphotino/photino.NET/issues/85
-            
-            var shortcutName = Path.ChangeExtension(AppEnvironment.ApplicationMainWindowTitle, "lnk");
-            var shortcutPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData, Environment.SpecialFolderOption.DoNotVerify), @"Microsoft\Windows\Start Menu\Programs", shortcutName);
-           
-            if (File.Exists(shortcutPath))
+
+            // Remove only if per-machine, don't delete the shortcut if the current instance of the app is portable or per-user
+            if (AppEnvironment.IsInstalledPerMachineAppInstance)
             {
-                try
+                var shortcutName = Path.ChangeExtension(AppEnvironment.ApplicationMainWindowTitle, "lnk");
+                var shortcutPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData, Environment.SpecialFolderOption.DoNotVerify), @"Microsoft\Windows\Start Menu\Programs", shortcutName);
+
+                if (File.Exists(shortcutPath))
                 {
-                    File.Delete(shortcutPath);
-                }
-                catch (IOException)
-                {
-                    // ignore "The process cannot access the file '..\Bravo for Power BI.lnk' because it is being used by another process."
+                    try
+                    {
+                        File.Delete(shortcutPath);
+                    }
+                    catch (IOException)
+                    {
+                        // ignore "The process cannot access the file '..\Bravo for Power BI.lnk' because it is being used by another process."
+                    }
                 }
             }
         }
